+++ src/astro/sql/operators/sql_decorator.py
@@ -26,7 +26,7 @@ from airflow.providers.snowflake.hooks.snowflake import SnowflakeHook
 from airflow.utils.db import provide_session
 from sqlalchemy.sql.functions import Function
 
-from astro.sql.table import Table
+from astro.sql.table import Table, create_table_name
 from astro.utils import postgres_transform, snowflake_transform
 from astro.utils.load_dataframe import move_dataframe_to_sql
 from astro.utils.schema_util import set_schema_query
@@ -118,7 +118,7 @@ class SqlDecoratoratedOperator(DecoratedOperator):
             self._set_schema_if_needed()
 
             if not self.output_table:
-                output_table_name = self.create_table_name(self.schema_id, context)
+                output_table_name = create_table_name(self.schema_id, context)
             else:
                 output_table_name = self.output_table.qualified_name()
             self.sql = self.create_temporary_table(self.sql, output_table_name)
@@ -140,10 +140,6 @@ class SqlDecoratoratedOperator(DecoratedOperator):
 
         if self.output_table:
             return self.output_table
-        #     sql = self.create_temporary_table(
-        #         f"-- SELECT * FROM {output_table_name}", self.output_table.table_name
-        #     )
-        #     self._run_sql(sql, self.parameters)
 
         elif self.raw_sql:
             return query_result
@@ -257,12 +253,6 @@ class SqlDecoratoratedOperator(DecoratedOperator):
     @staticmethod
     def create_cte(query, table_name):
         return f"WITH {table_name} AS ({query}) SELECT * FROM {table_name};"
-
-    @staticmethod
-    def create_table_name(schema_id, context):
-        ti: TaskInstance = context["ti"]
-        dag_run: DagRun = ti.get_dagrun()
-        return f"{schema_id}.{dag_run.dag_id}_{ti.task_id}_{dag_run.id}"
 
     @staticmethod
     def create_output_csv_path(context):
